# contracts/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Lrail API
  version: 1.0.0
  description: >-
    Organization-scoped API for Lrail projects, services, deployments,
    environments, logs and domains. Resource identifiers are opaque.
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /services/{service_id}/deployments:
    post:
      tags: [Deployments]
      summary: Create an immutable deployment
      operationId: createDeployment
      parameters:
        - $ref: "#/components/parameters/ServiceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeploymentRequest"
      responses:
        "202":
          description: Deployment accepted
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        default:
          $ref: "#/components/responses/Error"
...
  /deployments/{deployment_id}/cancel:
    post:
      tags: [Deployments]
      summary: Cancel a non-terminal deployment
      operationId: cancelDeployment
      parameters:
        - $ref: "#/components/parameters/DeploymentId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelDeploymentRequest"
      responses:
        "202":
          description: Cancellation accepted
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  headers:
    CorrelationId:
      description: Stable identifier for tracing this request across components.
      schema:
        type: string
        format: uuid
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Stable caller-generated key for one logical mutation.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
    ServiceId:
      name: service_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/PublicId"
    DeploymentId:
      name: deployment_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/PublicId"
...
  responses:
    Error:
      description: Request failed
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    PublicId:
      type: string
      description: Opaque, non-sequential resource identifier.
      minLength: 8
      maxLength: 64
      pattern: "^[A-Za-z0-9][A-Za-z0-9_-]+$"
...
    Deployment:
      type: object
      required: [id, organization_id, service_id, status, source, created_at]
      properties:
        id:
          $ref: "#/components/schemas/PublicId"
        organization_id:
          $ref: "#/components/schemas/PublicId"
        service_id:
          $ref: "#/components/schemas/PublicId"
        environment_id:
          $ref: "#/components/schemas/PublicId"
        revision_id:
          type: string
          minLength: 8
          maxLength: 64
          pattern: "^[A-Za-z0-9][A-Za-z0-9_-]+$"
          nullable: true
        status:
          type: string
          enum:
            - created
            - queued
            - preparing
            - building
            - scanning
            - deploying
            - verifying
            - ready
            - promoted
            - superseded
            - canceling
            - canceled
            - failed
        source:
          $ref: "#/components/schemas/DeploymentSourceSnapshot"
        preview_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Operation:
      type: object
      required: [id, organization_id, status, correlation_id, created_at]
      properties:
        id:
          $ref: "#/components/schemas/PublicId"
        organization_id:
          $ref: "#/components/schemas/PublicId"
        resource_id:
          $ref: "#/components/schemas/PublicId"
        status:
          type: string
          enum: [pending, running, succeeded, failed, canceled]
        correlation_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [code, message, correlation_id]
      properties:
        code:
          type: string
          pattern: "^[a-z][a-z0-9_]*$"
        message:
          type: string
        correlation_id:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
...
    DeploymentSourceSnapshot:
      oneOf:
        - type: object
          additionalProperties: false
          required: [type, reference, repository_id, commit_sha]
          properties:
            type:
              type: string
              enum: [git]
            reference:
              type: string
              minLength: 1
            repository_id:
              type: string
              minLength: 1
            commit_sha:
              type: string
              pattern: "^[0-9a-f]{40,64}$"
            root_directory:
              type: string
              minLength: 1
              maxLength: 1024
        - type: object
          additionalProperties: false
          required: [type, reference, digest]
          properties:
            type:
              type: string
              enum: [oci]
            reference:
              type: string
              minLength: 1
            digest:
              type: string
              pattern: "^sha256:[0-9a-f]{64}$"
...
    CreateDeploymentRequest:
      type: object
      required: [source]
      properties:
        source:
          $ref: "#/components/schemas/DeploymentSourceSnapshot"
        environment_id:
          $ref: "#/components/schemas/PublicId"
    CancelDeploymentRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
...